+++ packages/harmony_auth/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.1.0
 
 - Added unionOf and intersectionOf implementations for AuthMatcher
+- Added extra field to AuthToken
 
 ## 1.0.0